MarkdownEditor: refactor
